requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
kiwisolver==1.4.5
matplotlib==3.8.3
minio==7.2.7
mpmath==1.3.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
torch~=2.2.2
torchvision~=0.17.2
pandas==2.2.2
scipy==1.13.0
tensorboardX==2.6.2.2
tqdm==4.66.4
flask
pymongo
pydantic
python-dotenv
sqlalchemy
celery
redis
psycopg2-binary
minio
flask-login
flask-cors
cryptography
alembic
flower
